+++ tests/requests/invalid/002.py
@@ -1,2 +1,2 @@
-from gunicorn.http.errors import InvalidRequestLine
+from gunicorn.ghttp.errors import InvalidRequestLine
 request = InvalidRequestLine
