+++ tests/requests/invalid/004.py
@@ -1,2 +1,2 @@
-from gunicorn.http.errors import InvalidHTTPVersion
+from gunicorn.ghttp.errors import InvalidHTTPVersion
 request = InvalidHTTPVersion
