+++ tests/requests/invalid/008.py
@@ -1,2 +1,2 @@
-from gunicorn.http.errors import LimitRequestHeaders
+from gunicorn.ghttp.errors import LimitRequestHeaders
 request = LimitRequestHeaders
